Bugfix: Allow args to override default page number
By default, calling where sets the page property to 1. If the user tries
to override the value for page by specifying { page: 2, pageSize: 10 }
for args, then the pageSize will be set correctly, but the value for
page will be overwritten when the call to `merge()` applies the value
for page, which is hard-coded to 1.